+++ CHANGELOG.md
@@ -119,6 +119,11 @@
   [Marcelo Fabri](https://github.com/marcelofabri)
   [#973](https://github.com/realm/SwiftLint/issues/973)
 
+* Add `unused_optional_binding` rule that will check for optional bindings 
+  not being used.  
+  [Rafael Machado](https://github.com/rakaramos/)
+  [#1116](https://github.com/realm/SwiftLint/issues/1116)
+
 ##### Bug Fixes
 
 * Ignore close parentheses on `vertical_parameter_alignment` rule.